requirements.txt
Flask>=3.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
anthropic>=0.18.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.0.0
gunicorn>=21.0.0
